Fixed issue with chat command listener
Not sure when it happened but somehow the default channel was no
longer being set to 75 thus it was listening on the default value
for an int (ie 0) now chatChannel is declared to be 75 directly in
the definition.

This one explains some weird behaviour I was finding with the script
time that I couldn't narrow down before.