Few bugfixes
Fixed a few sources that had got out of sync with includes last commit

Also included the fix for #46

Was also getting a script failure in transform due to the new help
notecard it was trying to list as a button, due to this being an issue
that has come up before when we added notecards to the key I have
changed the way Transform.lsl decides a notecard is a Transforming
nc.  Rather than it being anything not otherwise blacklisted, now
any Transform notecard has a name starting with the prefix * this
was choosen for constantancy with the Dress.lsl implementation.