Bugfix and performance improvements on startup.
Several duplicate link messages and other minor startup niggles have
been cleaned up improving startup time.

Much larger gains in startup time have been added for script resets
by avoiding redundant reads of the preferences notecard.  Redundancy
being defined as all of the following being true.
- The HTTPdb is online and a record retreval search returns status
  200.
- There exists a record in the reply with key ncPrefsLoadedUUID.
- The record for ncPrefsLoadedUUID is not null and matches the UUID
  of the preferences notecard located inside the worn key.

A bug was found in Start.lsl which would result in the startup hanging
if there was no Preferences notecard available, now the startup will
simply give a debug message then jump to the doneConfiguration() step.
The same occurs where the notecard is not to be read because it is not
necessary.

Performance results:
- Script reset load times now under 5 seconds.
- On rez load times are now around 1 second.