Documentation of outfit folders and #RLV/ added
Documentation is being improved. Progressive updates will be
forthcoming and should be examed for errors.

The support for "Dressup" "> Dressup" and "Outfits" folders is
not going to be documented "directly" to encourage people to use
the "> Outfits" folder instead. The rest fall under the idea of
"legacy support" and may be documented there - perhaps in another
file.