Add some of he missing new commands to the chat command online help.
Also made a little fix to Menu Handler to not show sensor detection's to add
people already on the list.

Improved the name correction.

Abort query if add is specified for a user already on the list or the inverse
this is needed as unlike the menu the chat can take any username and we do
not want addRemBlacklist doing the wrong action.