Minor Several minor fixes
- MenuHandler gem light setter removed for some reason it is malfunctioning
  Start.lsl has a working one which now also handles run-time.
- Builtin controllers were locked out from carried dolls, this should not be
  the case as built-ins are a special situation.  It is recommended as in all
  cases that such access is used sparingly and if the carrying avatar is not
  afk it is always better to simply ask if they could place them down for a
  moment.
- Some fixes added the the sensor results dialog when adding a controller or
  a blacklist user via the menu.  Options will not be included for any user
  who is already listed.

Main
- Small improvement which was lost in earlier revert,

Start
- Reduced repetition in process configuration grouping the majority of configs
  with internal names from their noMinor Several minor fixes

- MenuHandler gem light setter removed for some reason it is malfunctioning
  Start.lsl has a working one which now also handles run-time.
- Builtin controllers were locked out from carried dolls, this should not be
  the case as built-ins are a special situation.  It is recommended as in all
  cases that such access is used sparingly and if the carrying avatar is not
  afk it is always better to simply ask if they could place them down for a
  moment.
- Some fixes added the the sensor results dialog when adding a controller or
  a blacklist user via the menu.  Options will not be included for any user
  who is already listed.

Main
- Small improvement which was lost in earlier revert,

Start
- Reduced repetition in process configuration grouping the majority of configs
  with internal names from their notecard name, I think we should reconsider
  the notecard option nameing however.
- Added the ChatHandler script to the list of core scripts to monitor for
  health during startup.
- Change the gem lighting colour setup to attempt all three potential data
  sources in the following order:
     1. HTTPdb - Live source that conviently maintains the users preference
        even if the object changes.
     2. Prim description - Saved when the gem colour is updated so this is
        also a live value but not portable and can also suffer to SL tech
        failures.
     3. DataAppearance - This is used as a backup data source due to it's
        static nature, in addition some inspiration with notecards such as
        this one is taken from the OpenCollar project where collar designers
        can include a similar defaultsettings card.  In both cases it is by
        design the settings therein be considered defaults and should not
        overwrite changes made at run-time.
- Made some improvements to the dataserver handler for reading the prefs nc
  mostly formatting and a few small performance improvements.
- New feature added for documentation of configuration notecards to save the
  time penalty of reading notecard lines that do not contain data of value to
  the scripts.  The Preferences notecard handler now has two special tokens
  which can be used to direct the way in which the the notecard is read.

New control tokens explained

1. <NEXT-READ-LINE> - Skip ahead to specified lines
Format:
<NEXT-READ-LINE>#</NEXT-READ-LINE>
The parameter symbolised by the hash (#) may be one of the following:
An absolute line number where the first line in the notecard is 1 *NOT* 0
A plus (+) character followed by an integer number of lines to skip ahead.

Note: The purpose of this directive is to skip lines to speed up and as such
specifying a line above the line number where this token is read or using a
negative index is not supported nor is specifying the current line or any
index of 0.  Both conditions will result in an error message however allowing
this would not only defeat the intent it would also enable the potential an
inexperienced user or a typo could create a reading loop situation.

<END-OF-CONFIGURATION>
This token tells the key that there are no further key readable lines in the
current notecard and has an effect identical to reading the LSL constant
EOF from the notecard.  No futher read attempts will be made and any lines
bellow this point can be used freely for human readable explainations and/or
notes.

The skipped lines following a next read line token may also be used for free
text.  In both cases there is no need to mark such lines with comment markup
however you may prefer to do so anyway for readability or other reasons.

All tokens must be specified on their own otherwise blank line additional
text including comments are note supported.  If you need a comment it is
recommended to add it on the line that follows the directive which has the
additional bonus it saves a read for the key.
tecard name, I think we should reconsider
  the notecard option nameing however.
- Added the ChatHandler script to the list of core scripts to monitor for
  health during startup.
- Change the gem lighting colour setup to attempt all three potential data
  sources in the following order:
     1. HTTPdb - Live source that conviently maintains the users preference
        even if the object changes.
     2. Prim description - Saved when the gem colour is updated so this is
        also a live value but not portable and can also suffer to SL tech
        failures.
     3. DataAppearance - This is used as a backup data source due to it's
        static nature, in addition some inspiration with notecards such as
        this one is taken from the OpenCollar project where collar designers
        can include a similar defaultsettings card.  In both cases it is by
        design the settings therein be considered defaults and should not
        overwrite changes made at run-time.
- Made some improvements to the dataserver handler for reading the prefs nc
  mostly formatting and a few small performance improvements.
- New feature added for documentation of configuration notecards to save the
  time penalty of reading notecard lines that do not contain data of value to
  the scripts.  The Preferences notecard handler now has two special tokens
  which can be used to direct the way in which the the notecard is read.

New control tokens explained

1. <NEXT-READ-LINE> - Skip ahead to specified lines
Format:
<NEXT-READ-LINE>#</NEXT-READ-LINE>
The parameter symbolised by the hash (#) may be one of the following:
An absolute line number where the first line in the notecard is 1 *NOT* 0
A plus (+) character followed by an integer number of lines to skip ahead.

Note: The purpose of this directive is to skip lines to speed up and as such
specifying a line above the line number where this token is read or using a
negative index is not supported nor is specifying the current line or any
index of 0.  Both conditions will result in an error message however allowing
this would not only defeat the intent it would also enable the potential an
inexperienced user or a typo could create a reading loop situation.

<END-OF-CONFIGURATION>
This token tells the key that there are no further key readable lines in the
current notecard and has an effect identical to reading the LSL constant
EOF from the notecard.  No futher read attempts will be made and any lines
bellow this point can be used freely for human readable explainations and/or
notes.

The skipped lines following a next read line token may also be used for free
text.  In both cases there is no need to mark such lines with comment markup
however you may prefer to do so anyway for readability or other reasons.

All tokens must be specified on their own otherwise blank line additional
text including comments are note supported.  If you need a comment it is
recommended to add it on the line that follows the directive which has the
additional bonus it saves a read for the key.