Altered memory report.
The new version is intended to be less spammy for users without dev or test
compiled keys, the full report for every script is only show when explicitly
requested.  Such as in /75stats there is a limited startup memory report in
place however the data is filtered to only show those scripts which after
memory scaling are running with almost all their possible available memory
allocated.  This will usually be scripts with under 6kB of available memory
and highlighting only these it is more likely to draw the users attention
to issues.