Renamed Dress -> Outfits... added up one folder button
This renames the Dress button on Main menu to Outfits... for consistency it
also adds an up one folder option for easier browsing of RLV folders swapping
to main menu when in the highest allowable folder level.  This is possible
in Dress as it's stucture enforces a blocking menu on us anyway for now.