Local database storage of configuration
This adds an accessible internal configuration database to the key the main
imperative for this was to ensure that when a key is brought into online mode
that the remote database can be immediately initialised to to the current key
configuration.  Previously only changes would get stored which may have left
keys in some undefined state.

A link 301 message needs to be triggered when the key is changed back from
offline mode to online mode to cause all settings to be replayed and thus be
captured by ServiceRequestor and delivered to the database.