Fixes in Start.lsl
Couple of fixes and an enhancement for Start.lsl

Fixes:
ncPrefsLoadedUUID is now no longer a single UUID value but a list of recently
read UUID's the problem which occurs with storing only one happens when the
same avatar uses multiple seperate key objects for example because they have
one properly sized and adjusted for their petite avatar and one for biggie.

When the key is swapped it encounters it's old notecard again but it has now
been forgotten so the key reads it and overwrites the newer DB settings with
it this can be frustrating for the user especially if the notecard is older
by a large margin because they have not used the second form in a while.


Fix a bug that was causing the gem lightng to fail to be set correctly also
the key now loses it's magical glow and lighting effects when the doll is
expired as a visual representation of it the life being lost from it.