Many: TransformingKey removed & Gender added
TransformingKey was for when the Key came in two variants: a standard
Key and a TransformingKey. There is no reason for this now as there
is only one Key.

The Gender code was mostly in place but had a few places it needed
to be added. There was also no way to change it in the options; this
has been changed and continues to be worked on.