extensive rewrite of Outfits/Type folder searching
The reliance on @findfolder was removed in favor of @getinv:
this allows one single RLV call to search for multiple possible
Outfits folder names, and makes the RLV work in earlier versions.
A second RLV channel was added to separate the typeFolder search
from the outfitsFolder search. The start of the process was
shifted from the timer (which generated a mandatory 15s delay)
to an RLV call. A flag variable was added to keep multiple
processes from colliding with one another. The ~normalself
and ~nude folders are checked for when the Outfits search is
completed and the results propogated.

End result is a much faster folder searching process.