Minor adjustments to code
On startup with RLV check incomplete, key now is quieter and
sends message to agent, not world - and notifies Dolly of who
attempted to touch their key.

Initial Timer setting after a collapse processing is now 15s
instead of 0.22s or 1s as before.

Tweaked the Dolly down message also.