Rework of ifPermissions; prep for CheckRLV.lsl
Reworked ifPermissions extensively; removed testing (mostly) testing
for NoScript land: if we never drop TakeControl permissions, we
don't need to test - but is that wise? Made some local vars into
globals - suppoed to save space and be faster, but hard to say for
sure. Ran out of memory, so prepared to move RLV checks to their
own script.