windTimes removed and windTime used
Many changes: removed setting of multiple wind times, and all
related items. Converted to a single wind time - which means no
need for a "Wind..." button, and no need for a list of wind times,
and no having to scan and sort a list. Much code was changed and/or
removed.