made pluslist into plusList; access menu updated
Converted the variable pluslist to plusList to be in proper
format as is customary; reworked the Access... menu to
not allow adding to empty lists, and to allow others than
just dolly to see and use it properly (for example, Builtin
Controllers, and Controllers). Builtin Controllers have near
full access, and standard Controllers can remove themselves
or others from the Access list.