Remvoed Factory Reset, and updated help opts
Removed Factory Reset: it is at once too complicated and too easy.
Too complicated in that it required a large amount of code and
another listener and timer to boot. Too easy in that the option
is always present and always visible to the dolly as a temptation.
Will be replaced by a Safeword.