Added DEBUG_MODE and lmPluginSend() and more
DEBUG_MODE is now for debugging and is expected to be probably on
when DEVELOPER_MODE or TESTER_MODE is active. lmPluginSend is a
link message sender to send to plugins. The TESTER_MODE now is
precluded from being active when the DEVELOPER_MODE is - the
reverse should be true, but TESTER_MODE is set afterwards.