Extensive rework of Chat Listener
Many changes were made. The commands were examined to see if they
needed more code to adhere to things like canPose, canCarry, and
so on. The help command was pulled out to allow everyone access
to it, and the output adjusted based on the user (Doll, Controller,
Builtin Controller, or other) and DEVELOPER_MODe and TESTER_MODE and
DEBUG_MODE. The chat commands now return immediately on execution;
this prevents scanning all commands even if we're done. Many comments
were added. The chatPrefix can now be between 2 and 10 characters
long. The devhelp and listhelp commands were removed. The carry
and uncarry commands do more checking over who is calling them
exactly.