Fixed type access and access controls
The Builder type is ONLY available to the Dolly IF they are a
controller AND a Built-in Controller... The Key type, likewise.
The Slut type is ONLY available to the Dolly IF they are NOT
in ADULT Mode. All internal types will show - IF appropriate
without a notecard available - they are the ONLY ones to do this.