Bugfix: check for avatar in isDresser()
Before the code would put out a message with a missing user name
whenever the scripts called for an outfit programmatically.
The isDresser() function now checks for the presence of an
avatar with the id before printing the message, and the function
is not called for the internal code any more either.