Converted afk to lmSetConfig; rem setAFK int cmd
Removed internal command setAFK and made it into a new function
setAfk(afk). All setting of afk goes through lmSetAfk except the
menu which sends it and assumes... "converted" afk to tri-state,
at least in setting: setting afk to 2 is same as setting to 1,
except that 2 marks an automatic setting: autoAFK is set based
on this.