Many updates: converted can* to allow*
Renamed canPose, canDress, canCarry, and canRepeatWind to
allowPose, allowDress, allowCarry, and allowRepeatWind respectively:
these are settings that affect what the public is allowed to do.
Moved some code from MenuHandler to Transform to spread out
memory usage: both can now be considered fully using memory.
Hovertext workings moved to an internal command and macros moved
to include files. Text tweaked and bugs fixed. LANDMARK_CDROOM
renamed to LANDMARK_CDHOME (current home is not a "room" somewhere).
windRechargeTime moved to common globals. dollType is not "preset"
with a default; it is the setDollType routine that sets the default
during the startup process. setDollType now refuses to run if the
dollType requested is already set as current.