Reworked Outfits menu to use sep channel; fix bugs
New rework code required some bugfixes. Reworked code so that
the resulting Outfits menus would put the results on a different
channel OTHER than the dialogChannel - to prevent namespace
collisions (what if there was an outfit with the same name as
a Colour? or even an outfit named "240min" or "Options"?)