Revamped RLV processing drastically
Removed processing of RLV with buffered commands, saving commands
and Scripts using them into a list, and much more - and replaced
with a simple routine with less than half a dozen lines. The external
calls remain the same, and will for now. Used @getstatus to get
status of current restrictions and save them - to be refreshed
later.