Fixed clear & collapseMenu; name fix; baseRLV fix
THe RLV clear command was not being properly used, so this was
redone. The baseRLV was incorporated as part of this fix.
The extraneous clear commands were fixed or redone to accomodate
the new RLV clear command works. The collapseMenu was not being
triggered; fixed - as well as collapseTime being used in Transform.lsl
the same way as all the rest. The transformerId variable was renamed
to match others such as dollID (to transformerID).