Fixed multiple problems relating to random outfits
Random outfits were not being applied. Firstly, "feeding" the listener
a menu item turned out to be not doable: it ignores llSay from the
same prim. The selected outfit then was seen to always be number 0 -
this was fixed. Other fixes applied. Random outfits now generate
a "confirmation" menu with one outfit only - and this is used to
feed the listener.