Reduced Dress memory usage & removed unwear check
Dress script was running out of memory; the problem was narrowed
down to the number of outfits in a given directory. Memory usage
was reduced by eliminating outfitsList (by assigning to [] whenever
no longer needed). Also removed the checking for everything being
removed: the new outfit may contain elements of the old, such as
when an outfit comes with different length skirts or other similar
reasons.