Reworked memory reporting completely
The original memory reporting code was overly complex, and conflated
the Memory Report with the Memory Reply. The two processes were
separated, and the use of JSON removed.

Now the code builds a report in a list, then sorts the list and
concatenantes the list together in a string and outputs it to the
requestor.