Put in separate typeChannel and poseChannel
There was the potential for "crosstalk" between Poses, Types, and
Menu Selections: what if a Pose had the same name as a Menu Selection...
such as "Wind" for example? Bad things could happen - and same with
a Pose with the same name as a Type. The different channels fixes
this. (it was an extensive though important change).