Poses Menu fixed and reworked
The Poses Menu now uses a function to generate the actual menu page,
and uses a separate channel to keep poses from triggering Menu options.
The actual routine is called via InternalCommand in order to make
a new event for it. Page Next and Page Prev are implemented and
tested. The poseList now acts like a buffer, and is filled on
attach or Inventory Change - but also will be filled by the
page routine if need be (this might be overkill).