Poses fix for issue #431
The inventory count was made into a Global along the way, just like
the posesList. Turns out that while the count of Poses (or whatever)
is 1-indexed, the name retrieval is 0-indexed. This discrepancy
was causing grief. Also added a preload of the poseList and a
preload during attach and inventory change as well as a possible
load during pose paging.