Fixed a bug in outfits paging code
Somehow, a serious error returned to the outfits paging code.
Was most noticeable in directories with less than nine outfits:
They were numbered in reverse, and the right number of buttons
was shown with the "right" outfits attached - or none, as the
case may be. Fixed: counting down requires the use of the END
rather than the START index....