"Fixed" Glow and Light; optimized
The Key was not seen with any Glow or Light - and so "fixed" this
bug. Turned out that all I did was learn about Glow and Light - and
optimize - as Prim Glow and Prim Light didn't show up at the graphics
settings I was using. Removed the "ALL" shortcut for the notecard:
it gives very little value for the processing time it takes.