Rem need for JSON; fixed gemColour, light and glow
The primLight and primGlow did not seem to be working - by the time
it was discovered it was my graphics settings, I'd optimized the
hell out of the code :) The need for the DataAppearance notecard
and JSON has been removed. JSON is slower than normal LSL and is
of limited use unless there's some kind of JSON that has to be read
from somewhere.