Fixed afk and autoAfk problems
Issue #465: fixed a bug where Dolly would go out of AFK prematurely.
Reworked vars afk and autoAfk: afk means that Dolly is (TRUE) or
is not (FALSE) afk; autoAfk means that it is (TRUE) or is not
(FALSE) an automatically generated AFK. If afk is FALSE, then
autoAfk is meaningless and should be zero.