Reviewed Builder code to make dev only (#445)
The Builder Type code should only exist in very few places in a
normal Adult or Child Key: specifically, one should not be able
to use a notecard to become a Builder Dolly. In all other cases,
Builder code should be in Developer code only: others do not become
Builder Dollies, and testing for that type is a waste of cycles.