Converted times to integer (#486)
The more I looked at it the more I saw implicit conversions to
float all over the place - and the assumption was often that the
wind amounts or time left on key or key limit or other similar
were integers - when they were not. All should know be working
as integers.