Convert wind rate to percentage (issue #535)
This simplifies things drastically: the wind rate is a percentage
of the default 2𝜋 divided by 8 - one full rotation in 8 seconds.

Default could certainly be changed, but wind rate remains
a percentile of the default. Normal operation is 1, afk operation
is 0.5, and others as desired. Unwind operation could be set
briefly to 1.5 or 2.