Added UUID & name fixes (issue #578)
Controllers and blacklisted individuals are stored by UUID and by name.
Names cannot be gotten by a simple function call if the avatar specified
is not in the region. Getting the name from an avatar not in region
is more complex.

This code handles that, and removes name usage from Aux where it
is not needed.