Reworked ~normalself and ~nude
The ~normalself and ~nude folders are now separated for different
purposes. Removed "Reset Body" option - no longer needed - and set
~normalself to be that which is reset on each outfit change.

That is, ~normalself is standard and should always be worn.

Now, ~nude is only attached to complete a nude look - optional
items prone to be changed will be here: things like hair, eyes,
lipstick: things that might be changed out for stylistic reasons.

This resolves issue #579.