Lowered excessive use of windingDown
The configuration value windingDown seemed to be overly set and
propogated; changed to only be sent on change (issue #582).

Also changed collapse to not duplicate efforts to set the hovertext,
and to have a more simplified path. Now that the value collapsed
is once again binary, the excessive code is no longer needed.

Further testing is needed.