Fix autoAfk detection
Moved autoAfk code from Transform.lsl to Main.lsl. Frequency of
timing will not change, but setting will be set at same time
as the computation of current time remaining (getTimeUpdates).
This is as it should be.

This is related to issue #603 (though it doesn't fix the freq of
checking).