Changed reset sequence after update
The timing was off when the updater reset using the Start.lsl
script - mainly because the script had to restart all the other
scripts before things could happen.

With this fix (for issue #608) the Key will reset using New.lsl
after an update. The scripts will be started then Start.lsl will
be started, allowing it to reset all of the now running scripts.