Fixed problem with Outfits menu not showing
Outfits menu would not show; problem was with locking. Made
sure to clear lock (dresserID) on all exits from the dressing
flow, and also added a timer section (which may or may not have
existed previously) to provide a timeout.

Code showed a presumption of a timer section, but none existed -
was it deleted at one point?