Implemented and tested Reset Body
Reset Body button was moved to Help menu, tested, and a default
outfit added (from ~normaloutfit folder).

Outfit removal was expanded to include everything in the #RLV folder.
This allows one to wear an avatar, or extras, or anything else not
included in an outfit, but included in #RLV - and have it removed to
generate the default.

This is issue #672.