Completely new follow code
Per issue #702, the follow code was completely reworked. The code
was rewritten into a sort of pseudo-API, with standard functions
performing the work. The code itself was pulled out into an
include file, which allows a simple replacement of one method
for following for another.

The Follow.inc file uses llMoveTo and a timer to check on location
and to follow the carrier.

Follow2.inc will use the traditional llTarget and at_target /
not_at_target. The current version of this file does not work.

There could be others in the future.