Fixed error checking for missing poses
Per issue #713 and #714, added and fixed code to handle when poses
are missing, even while accounting for the collapse animation.

Further error checking for the collapse animation itself is needed;
this commit does not fully handle that.

Also added a few functions to be used in scripts that check for
poses and types.