Many updates + reworked initialization
CheckRLV.lsl and Start.lsl were changed so that the RLV check
is part of the initialization process, and as of Stage2 RLV has
been checked. This is a partial resolution of issue #717.

This commit also contains fixes for issue #615, #550, and #721.

There were also several minor updates to increase readability
and maintainability.