Reworked entire outfit searching process
Outfit searching was separated from type searching, and also
system folder searching was incorporated as part of outfit searching.

Outfit folders (and type folders) are now searched for AFTER
preferences file is completely read. Preferences settings
are picked up later after the file is read, and used to trigger
proper settings - or defaults, if no choice is specified.

The type "Regular" is default, and the outfits folder defaults
remain "> Outfits" or "> Dress" and related.

These fixes resolve issues #733, #734, and #735.