Enabled HOMING_DEVICE for testing
Per issue #667, enabled the homing device for testing.

The homing device should disable the TP Home button, and force
TP the Dolly back to the "Home" LM in their inventory. If this
does not exist, then the homing device should be disabled, as
well as the TP Home button.

Testing continues...