Reworked chat help, and controller test macros
Chat help needed to be reworked for permissions, and also examined
closely for genuine allowances: the commands in the help command
MUST be available on the same permissions.

Also increased readability in the Chat process.

Testing for controller status was changed and cleaned up, since
built-in controllers have been removed. Removed references to
"user controllers" and modified macros to delineate between
Dolly Controllers and Dolly.