Added button to re-check for RLV
If RLV check fails, then add a RLV check button in Help.
This is different than the optional RLV: once enabled, there
is no way to turn off RLV in the menu system.

The purpose is for the occasion when there is RLV, but the check
fails because of extended timeout. Using the button, RLV can be
enabled.

This is part of issue #766.