Tried to improve memory & speed in ChatHandler
Made code changes to make things run faster and with
fewer amounts of code.

This is related to issue #793.

Finally, after putzing around the edges, moved addController et al
and remController etc into Start to free memory.

Pays to remember that this is DEVELOPMENT memory; when Development
code is removed it'll have much more free memory.