Cleaned up macros for first character folders
Reworked names, fixed transforming name (to type name), and started
working on folder entry that starts with an = to act like an Avatar
switch while keeping key.

Using a single lead character prefix of = will be much easier than
a single directory for it all - and was already imagined in code
(presumably by another coder).