Commented back out with TP_HOME
Getting llRequestInventoryData() to work properly turned out to be
a collossal waste of time. Very angry and upset: commented feature
out with use of TP_HOME macro variable.

Maybe some day when I've the patience of a saint or the knowledge
of one of the gods I can return to this.