Fixed bug with no type folder found messages
When a type folder is not found, this could be an error, or it could be
normal - or it could be expected.

When the type folder search times out, this is an error.

When the type folder is searched for, and not found: this requires a
message to let the user know.

If there is no type folder search - as none will be recognized - then
there should be no message to alarm the user.